Move graph tests
